+++ CHANGELOG.md
@@ -11,3 +11,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Harmony patch to inject new menu button triggers.
 - Addes short and medium pause options.
 - Added dual press, multi tap and button+trigger pause options.
+- Allow more customization when configuring menu button triggers.
+- Replaced menu button triggers with a new configurable menu button trigger.
+- Changed the harmony patch to make use of the configurable menu button trigger.